test_arize_callback
