+++ tests/local_testing/test_arize_ai.py
@@ -1,7 +1,10 @@
 import asyncio
 import json
 import logging
-
+import os
+import time
+from unittest.mock import patch, Mock
+import opentelemetry.exporter.otlp.proto.grpc.trace_exporter
 from litellm import Choices
 import pytest
 from dotenv import load_dotenv
@@ -60,3 +63,49 @@ def test_get_arize_config_with_endpoints(mock_env_vars, monkeypatch):
     config = ArizeLogger.get_arize_config()
     assert config.endpoint == "grpc://test.endpoint"
     assert config.protocol == "otlp_grpc"
+
+
+def test_arize_callback():
+    litellm.callbacks = ["arize"]
+    os.environ["ARIZE_SPACE_KEY"] = "test_space_key"
+    os.environ["ARIZE_API_KEY"] = "test_api_key"
+    os.environ["ARIZE_ENDPOINT"] = "https://otlp.arize.com/v1"
+
+    # Set the batch span processor to quickly flush after a span has been added
+    # This is to ensure that the span is exported before the test ends
+    os.environ["OTEL_BSP_MAX_QUEUE_SIZE"] = "1"
+    os.environ["OTEL_BSP_MAX_EXPORT_BATCH_SIZE"] = "1"
+    os.environ["OTEL_BSP_SCHEDULE_DELAY_MILLIS"] = "1"
+    os.environ["OTEL_BSP_EXPORT_TIMEOUT_MILLIS"] = "5"
+
+    try:
+        with patch.object(
+            opentelemetry.exporter.otlp.proto.grpc.trace_exporter.OTLPSpanExporter,
+            "export",
+            new=Mock(),
+        ) as patched_export:
+            litellm.completion(
+                model="openai/test-model",
+                messages=[{"role": "user", "content": "arize test content"}],
+                stream=False,
+                mock_response="hello there!",
+            )
+
+            time.sleep(1)  # Wait for the batch span processor to flush
+            assert patched_export.called
+    finally:
+        # Clean up environment variables
+        for key in [
+            "ARIZE_SPACE_KEY",
+            "ARIZE_API_KEY",
+            "ARIZE_ENDPOINT",
+            "OTEL_BSP_MAX_QUEUE_SIZE",
+            "OTEL_BSP_MAX_EXPORT_BATCH_SIZE",
+            "OTEL_BSP_SCHEDULE_DELAY_MILLIS",
+            "OTEL_BSP_EXPORT_TIMEOUT_MILLIS",
+        ]:
+            if key in os.environ:
+                del os.environ[key]
+
+        # Reset callbacks
+        litellm.callbacks = []
